requirements.txt
playwright
pytest
allure-pytest
pyyaml
colorama
pytest-check
